serveur/requirements.txt
django==5.0.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.*
django-cors-headers==4.*
django-filter==24.*
django-ratelimit==4.*
python-decouple==3.*
drf-spectacular==0.27.*
celery==5.*
redis==5.*
flower==2.*
psycopg2-binary==2.*
weasyprint==62.*
qrcode[pil]==7.*
Pillow==10.*
# Bibliothèques métier & Sécurité
web3>=6.20.0
pytest==8.*
pytest-django==4.*
factory-boy==3.*
faker==24.*
ruff==0.4.*
black==24.*
sentry-sdk==2.*
